+++ rivershared/util/testutil/test_util.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"time"
 )
 
 // See docs on PanicTB.
@@ -118,4 +119,23 @@ type TestingTB interface {
 	Log(args ...any)
 	Logf(format string, args ...any)
 	Name() string
+}
+
+// WaitTimeout returns a duration broadly appropriate for waiting on an expected
+// event in a test, and which is used for `TestSignal.WaitOrTimeout` in
+// testsignal and `WaitOrTimeout` in riversharedtest. Its main purpose is to
+// allow a little extra leeway in GitHub Actions where we occasionally seem to
+// observe subpar performance which leads to timeouts and test intermittency,
+// while still keeping a tight a timeout for local test runs where this is never
+// a problem.
+//
+// It lives here instead of riversharedtest so that testsignal, which is
+// compiled into production binaries, can use it without importing
+// riversharedtest and its heavier test-only dependencies.
+func WaitTimeout() time.Duration {
+	if os.Getenv("GITHUB_ACTIONS") == "true" {
+		return 10 * time.Second
+	}
+
+	return 3 * time.Second
 }